replace with rhythm_reducer.lua
"rest_minimiser" is in the trash and the new version is renamed "rhythm_reducer.lua" using `FCNoteEntryCell::ReduceEntries()`. It feels like a cheap trick but at least this gives easy access to the routine built into RGP0.68+!
